+++ src/components/Icon/Icon.stories.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import viewsPalette from 'theme/palette';
+import palette from 'theme/palette';
 import Icon from './Icon';
 import icons from './icons.json';
 
 const control = {
 	type: 'select',
-	options: Object.keys(viewsPalette).reduce((options, colorName) => {
-		options[colorName] = viewsPalette[colorName];
+	options: Object.keys(palette).reduce((options, colorName) => {
+		options[colorName] = palette[colorName];
 		return options;
 	}, {})
 };