+++ src/components/Skeleton/styles.js
@@ -1,12 +1,12 @@
 import styled from 'styled-components';
-import viewsPalette from 'theme/palette';
+import palette from 'theme/palette';
 
 export default {
 	SkeletonContainer: styled.div`
 		border-radius: ${({ circle }) => (circle ? ' 50%' : '3px')};
 		height: ${({ height }) => height};
 		width: ${({ width }) => width};
-		background-color: ${({ backgroundColor }) => backgroundColor || viewsPalette.lightGrey};
+		background-color: ${({ backgroundColor }) => backgroundColor || palette.lightGrey};
 		animation: pulse 1.5s ease-in-out infinite;
 
 		@keyframes pulse {