+++ src/theme/palette.js
@@ -1,4 +1,4 @@
-const viewsPalette = {
+const palette = {
 	black: '#2F2F2F',
 	blackHover: '#585858',
 	blackPressed: '#16232D',
@@ -7,6 +7,7 @@ const viewsPalette = {
 	blueHover: '#5393FF',
 	bluePressed: '#1759FF',
 	blueAvailable: '#E4ECFA',
+	darkBlue: '#0016DD',
 	darkGrey: '#939598',
 	darkGreyHover: '#A8AAAC',
 	darkGreyPressed: '#747679',
@@ -26,7 +27,8 @@ const viewsPalette = {
 	lightGrey: '#E8EAF6',
 	lightGreyHover: '#F4F5FB',
 	lightGreyPressed: '#D0D3E3',
-	lightGreySelected: '#E4ECFA',
+	lightTurquoise: '#BBE9D6',
+	magenta: '#ED14A4',
 	orange: '#FF8D10',
 	orangeHover: '#FFA33F',
 	orangeLight: '#ffead3',
@@ -50,4 +52,4 @@ const viewsPalette = {
 	yellowPressed: '#FFBA0C'
 };
 
-export default viewsPalette;
+export default palette;